adicionado confirmação de deletamento de notas

+++ src/AnotationBlock.jsx
@@ -1,6 +1,8 @@
 import './anotationBlock.css'
 import notasIcon from './assets/notinhasIcone.png'
 import apagarNota from './assets/apagarnota.png'
+import confirmar from './assets/confirma.png'
+import negar from './assets/negar.png'
 import { useState } from 'react'
 import {motion, stagger} from 'framer-motion'
 
@@ -9,19 +11,31 @@ function AnotationBlock({cor, titulo, subtitulo, notaID, refresh, animate}){
 
     const [confirma, setConfirma] = useState(false)
 
-    const deletarNota = async (notaID) => {
-        await fetch(`http://localhost:3333/controle/deletarnota/${notaID}`, {
+
+
+
+
+    const deletarNota = async (notaID, confirmacao) => {
+        setConfirma(true)
+        if (confirmacao == true){
+            await fetch(`http://localhost:3333/controle/deletarnota/${notaID}`, {
             method: 'DELETE',
             headers: {
                 'Content-Type': 'application/json'
             }
         })
         refresh()
-        
+        }
+        else{
+            setConfirma(false)
+        }
     }
 
 
     return(
+        
+
+
         <motion.div className='itemListaNota' style={{backgroundColor: cor}}
         variants={animate? {
             before: {marginLeft:"-250%"},
@@ -31,6 +45,7 @@ function AnotationBlock({cor, titulo, subtitulo, notaID, refresh, animate}){
         animate="after"
         exit="before"
         > 
+        
             <div style={{height:"100%", display:"flex", alignItems:"start", justifyContent:"center", padding:5}}>
                             <img className='notinhaIcon' src={notasIcon}></img>
             </div>
@@ -41,8 +56,16 @@ function AnotationBlock({cor, titulo, subtitulo, notaID, refresh, animate}){
             </div>
 
             <div style={{height:"100%", display:"flex", alignItems:"start", justifyContent:"center", padding:10}}>
-                <img onClick={() => deletarNota(notaID)} style={{height:50, width:50}} src={apagarNota}></img>
+                {!confirma && <img onClick={() => setConfirma(true)} style={{height:50, width:50}} src={apagarNota}></img>}
             </div>
+
+
+        {confirma && <div className='confirma'>
+            <div onClick={() => deletarNota(notaID, true)} className='flex-center' style={{backgroundColor:"red", height:"100%", width:"50%"}}><img src={confirmar}></img></div>
+            <div onClick={() => deletarNota(notaID, false)} className='flex-center' style={{backgroundColor:"green", height:"100%", width:"50%"}}><img src={negar}></img></div>
+        </div>}
+
+        {/* <button style={{height:50, width:50}} onClick={() => (setConfirma(c => !c))}></button> */}
         </motion.div>
     )
 }
